# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/notifications": {
      "get": {
        "operationId": "listNotifications",
        "parameters": [
          {
            "name": "unreadOnly",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationDto_Output"
                  }
                }
              }
            }
          },
          "400": {
            "description": "`VALIDATION_ERROR`: `unreadOnly` no es `true` ni `false`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED`: falta la cookie de sesión, está alterada o venció.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "notifications"
        ],
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/notifications/{id}/read": {
      "post": {
        "operationId": "markNotificationRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "La notificación, con `readAt` ya seteado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED`: falta la cookie de sesión, está alterada o venció.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "`NOT_FOUND`: la notificación no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "notifications"
        ],
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/notifications/read-all": {
      "post": {
        "operationId": "markAllNotificationsRead",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Cantidad de notificaciones no leídas que se marcaron como leídas.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkAllNotificationsReadDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED`: falta la cookie de sesión, está alterada o venció.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "notifications"
        ],
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "AI Trend Debates API",
    "description": "Pipeline automatizado que convierte un trend en un debate verificado entre agentes de IA.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "atd_session",
        "description": "Cookie httpOnly que emite POST /auth/login (7 días)."
      }
    },
    "schemas": {
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "VALIDATION_ERROR",
          "INVALID_SEQUENCE_INDEX",
          "UNAUTHORIZED",
          "INVALID_CREDENTIALS",
          "FORBIDDEN",
          "NOT_FOUND",
          "INVALID_STATE_TRANSITION",
          "MANIFEST_NOT_READY",
          "VOICE_NOT_CONFIGURED",
          "USAGE_LIMIT_EXCEEDED",
          "TOO_MANY_ATTEMPTS",
          "LOGIN_BUSY",
          "PROVIDER_QUOTA_EXCEEDED",
          "HTTP_ERROR",
          "INTERNAL_ERROR"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "$ref": "#/components/schemas/ErrorCode"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "NotificationDto_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "episodeId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "type": {
            "type": "string",
            "enum": [
              "EPISODE_COMPLETED",
              "EPISODE_PENDING_REVIEW",
              "EPISODE_REQUIRES_REVIEW",
              "EPISODE_FAILED"
            ]
          },
          "message": {
            "type": "string"
          },
          "readAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          }
        },
        "required": [
          "id",
          "episodeId",
          "type",
          "message",
          "readAt",
          "createdAt"
        ],
        "additionalProperties": false
      },
      "MarkAllNotificationsReadDto_Output": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "count"
        ],
        "additionalProperties": false
      },
...
    }
  }
}